# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Organization Management Service",
    "version": "1.0.0",
    "description": "Multi-tenant organization management API"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/org/create": {
      "post": {
        "summary": "Create a new organization",
        "tags": ["Organizations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organization_name": { "type": "string", "minLength": 2 },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 }
                },
                "required": ["organization_name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "organization_name": { "type": "string" },
                    "db_name": { "type": "string" },
                    "admin_email": { "type": "string" },
                    "created_at": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request / validation error" },
          "409": { "description": "Organization already exists" }
        }
      }
    },
    "/org/get": {
      "get": {
        "summary": "Get organization metadata",
        "tags": ["Organizations"],
        "parameters": [
          {
            "name": "organization_name",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Organization details" },
          "404": { "description": "Organization not found" }
        }
      }
    },
    "/org/update": {
      "put": {
        "summary": "Update organization admin credentials",
        "tags": ["Organizations"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organization_name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 }
                },
                "required": ["organization_name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Organization updated" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/org/delete": {
      "delete": {
        "summary": "Delete organization and its database",
        "tags": ["Organizations"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "organization_name",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Organization deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/admin/login": {
      "post": {
        "summary": "Authenticate admin and get tokens",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" },
                    "token_type": { "type": "string" },
                    "expires_in": { "type": "number" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Get new access token from refresh token",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "admin_id": { "type": "string" },
                  "refresh_token": { "type": "string" }
                },
                "required": ["admin_id", "refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token refreshed" },
          "401": { "description": "Invalid refresh token" }
        }
      }
    },
    "/admin/verify-token": {
      "get": {
        "summary": "Verify JWT token validity",
        "tags": ["Authentication"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Token is valid" },
          "401": { "description": "Invalid or expired token" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}